move serve_configuration to dispatch_configuration.rs
